+++ compottie/src/commonMain/kotlin/io/github/alexzhirkevich/compottie/internal/content/DrawingContent.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.graphics.Canvas
 import androidx.compose.ui.graphics.Matrix
 
 internal interface DrawingContent : Content {
-    fun draw(canvas: Canvas, parentMatrix: Matrix, parentAlpha : Float, frame: Int)
+    fun draw(canvas: Canvas, parentMatrix: Matrix, parentAlpha: Float, frame: Float)
 
-    fun getBounds(outBounds : MutableRect, parentMatrix: Matrix, applyParents: Boolean, frame: Int)
+    fun getBounds(outBounds: MutableRect, parentMatrix: Matrix, applyParents: Boolean, frame: Float)
 }
