+++ compottie/src/commonMain/kotlin/io/github/alexzhirkevich/compottie/internal/content/PathContent.kt
@@ -3,5 +3,5 @@ package io.github.alexzhirkevich.compottie.internal.content
 import androidx.compose.ui.graphics.Path
 
 internal interface PathContent : Content {
-    fun getPath(frame: Int) : Path
+    fun getPath(frame: Float) : Path
 }
