+++ compottie/src/commonMain/kotlin/io/github/alexzhirkevich/compottie/internal/content/ShapeModifierContent.kt
@@ -1,16 +1,14 @@
 package io.github.alexzhirkevich.compottie.internal.content
 
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.util.fastForEachReversed
-import io.github.alexzhirkevich.compottie.internal.schema.helpers.Bezier
 import io.github.alexzhirkevich.compottie.internal.schema.helpers.ShapeData
 
 internal interface ShapeModifierContent : Content {
-    fun modify(shape: ShapeData, frame : Int) : ShapeData
+    fun modify(shape: ShapeData, frame: Float) : ShapeData
 }
 
 
-internal fun ShapeData.modifiedBy(modifiers : List<ShapeModifierContent>, frame: Int) : ShapeData {
+internal fun ShapeData.modifiedBy(modifiers: List<ShapeModifierContent>, frame: Float) : ShapeData {
     var b = this
 
     modifiers.fastForEachReversed {
